chore: rename GeminiLessonService to ClaudeLessonService, remove debug logs, fix Long/Integer type

+++ backend/src/main/java/com/predicto/academy/UserLessonProgressRepository.java
@@ -20,5 +20,5 @@ public interface UserLessonProgressRepository extends JpaRepository<UserLessonPr
     List<UUID> findCompletedLessonIdsByUserId(@Param("userId") UUID userId);
 
     @Query("SELECT COALESCE(SUM(p.xpEarned), 0) FROM UserLessonProgress p WHERE p.user.id = :userId")
-    Integer sumXpEarnedByUser(@Param("userId") UUID userId);
+    Long sumXpEarnedByUser(@Param("userId") UUID userId);
 }
